Working toward provisioning, fix a buncha skipped tests
So yeah, if you misspell a constant name, and then "if that constant is present and has a crafting_type, then run this test", that test won't be run, because that constant isn't actually present if misspelled. It's nil.

So when you start refactoring and make testing code more generic so that it falls back on a different, but correctly spelled, field in the row to find a crafting_type? All of a sudden we're testing ebony and homespun items. And failing because they're spelled differently in condition text than in crafting table tooltips. Okay. Fixed.

Meanwhile, stop calling stuff "gear". The writ board calls BS/CL/JW/WW "equipment". And you'll need some way to differentiate from AL/EN/PR "consumables".