alch/: subproject to learn each alchemy traits on each reagent
Copy the little sub-project I ran to figure out the potion list that `/lct alchemist` creates.

Yeah, this has a repo of its own elsewhere, but I have a feeling that directory is not long for this world. Copying the code to here retains it for Future Zig.

Intentionally not copying the git history for this subproject. Nothing in there worth saving. Certainly nothing worth the hassle of a submodule.