Sets, Motifs added to lang/xx.lua
Several hundred of the strings in these lang files could be generated on demand, using the exact same code that we use for /lct discover.

But is the memory savings worth it? A few hundred strings is 100KB or less. I'd rather burn the strings into tables so that I can see the data at rest, rather than have to poke and peek at the code while running.